+++ tools/capi-print-bridge/src/preload.js
@@ -1,0 +1,5 @@
+﻿const { contextBridge, ipcRenderer } = require("electron");
+
+contextBridge.exposeInMainWorld("bridge", {
+  onLog: (callback) => ipcRenderer.on("bridge-log", (_event, lines) => callback(lines)),
+});